requirements.txt
# DeepStream Python bindings and core dependencies
# Note: DeepStream Python bindings (pyds) are typically installed with DeepStream SDK
# These are additional Python packages needed for the application

# Configuration management
PyYAML>=6.0
pydantic>=2.0

# Event delivery
requests>=2.28.0
paho-mqtt>=1.6.0

# File monitoring for hot-reload
watchdog>=3.0.0

# Image processing for frame saving
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
python-dateutil>=2.8.0
uuid

# Logging and monitoring
structlog>=23.0.0

# Web server and API
Flask>=2.3.0
Flask-CORS>=4.0.0

# System monitoring
psutil>=5.9.0
pynvml>=11.5.0

# AWS Kinesis Video Streams
boto3>=1.26.0
botocore>=1.29.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0